requirements.txt
langchain-community
langchain-google-genai
faiss-cpu 
pypdf
python-multipart
pandas
fastapi
uvicorn
pydantic
passlib
python-jose
bcrypt==4.0.1
python-pptx
networkx
openpyxl
docx2txt
docx
unstructured
